requirements.txt
Click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
pytest==6.2.2
Telethon == 1.25.2
cryptography == 38.0.1
WTForms == 3.0.1
flask-login == 0.6.2
flask-WTF == 1.0.1
setuptools==65.4.1
wheel==0.37.1
oauthlib==3.0.1
pyOpenSSL==22.1.0
requests==2.21.0
Flask-Migrate==2.5.2
pillow==9.2.0